requirements.txt
fastapi
uvicorn
requests

#for template
jinja2

#for static files
aiofiles

#for database   #new
sqlalchemy
psycopg2

#for loading environment variables  #new
python-dotenv